Arlie Jump Update + GroundSetter
Arlie can now be controlled in air:
- Hold Jump to jump higher!
- Let go of Jump to stop ascending!
- Short-Tap Jump for a mini-jump!
- Press Crouch in Air to descent fast!

The App-Window can now freely be resized without it causing Arlie to be unable to jump or be set to the top of the window.
This was achieved using another observer, observer the scene.getHeight() at any given time, setting the Ground at always *.7-Arlie's height as well as Arlie on top of it.